Move examples to integration tests. They can be used as examples too.
