+++ README.md
@@ -22,7 +22,7 @@ cat test/stringified-mdl.txt | cargo run -- get-namespaces -
 ## Library
 
 Here are some examples of how to use the library. You can see more in [examples](examples) folder and read about in the
-dedicated [README](examples/README.md).
+dedicated [README](tests/README.md).
 
 ### Eamples
 
@@ -66,8 +66,8 @@ sequenceDiagram
 4. **Reader Processing mDL data:**
     - The reader processes the response and prints the value of the `age_over_21` element.
 
-You can see the full example in [simulated_device_and_reader](examples/simulated_device_and_reader.rs) or a version that
-uses `State pattern`, `Arc` and `Mutex` [simulated_device_and_reader](examples/simulated_device_and_reader_state.rs).
+You can see the full example in [simulated_device_and_reader](tests/simulated_device_and_reader.rs) or a version that
+uses `State pattern`, `Arc` and `Mutex` [simulated_device_and_reader](tests/simulated_device_and_reader_state.rs).
 
 ##### Device perspective
 
@@ -108,10 +108,10 @@ stateDiagram
     ReadyToRespond --> Reader: handle_response
 ```
 
-The reader is simulated in [common](examples/common.rs) module (you can find the code in [examples](examples)),
+The reader is simulated in [common](tests/common.rs) module (you can find the code in [examples](examples)),
 and we focus on the code from the
 device perspective.
-You can see the full example in [on_simulated_device](examples/on_simulated_device.rs).
+You can see the full example in [on_simulated_device](tests/on_simulated_device.rs).
 
 ##### Reader perspective
 
@@ -134,7 +134,7 @@ stateDiagram
     Reader --> Device: new_request
 ```
 
-Now the reader is simulated in [common](examples/common.rs) module (you can find the code in [examples](examples)),
+Now the reader is simulated in [common](tests/common.rs) module (you can find the code in [examples](examples)),
 and we focus on the code from the reader's perspective.
 The code is considerably shorter.
-You can see the full example in [on_simulated_reader](examples/on_simulated_reader.rs).
+You can see the full example in [on_simulated_reader](tests/on_simulated_reader.rs).
