+++ tests/common.rs
@@ -1,60 +1,34 @@
-mod common;
-
 use anyhow::{anyhow, Context, Result};
 use signature::{SignatureEncoding, Signer};
 use uuid::Uuid;
 
-use isomdl::definitions;
 use isomdl::definitions::device_engagement::{CentralClientMode, DeviceRetrievalMethods};
-use isomdl::definitions::device_request::DocType;
+use isomdl::definitions::device_request::{DataElements, DocType, Namespaces};
 use isomdl::definitions::helpers::NonEmptyMap;
-use isomdl::definitions::{BleOptions, DeviceRetrievalMethod};
+use isomdl::definitions::{self, BleOptions, DeviceRetrievalMethod};
 use isomdl::presentation::device::{Document, Documents, RequestedItems, SessionManagerEngaged};
-use isomdl::presentation::{device, Stringify};
+use isomdl::presentation::{device, reader, Stringify};
 
-use crate::common::{establish_reader_session, reader_handle_device_response};
-
+#[allow(dead_code)]
 const DOC_TYPE: &str = "org.iso.18013.5.1.mDL";
 const NAMESPACE: &str = "org.iso.18013.5.1";
 const AGE_OVER_21_ELEMENT: &str = "age_over_21";
 
-fn main() -> Result<()> {
-    run_on_simulated_device()
-}
-
-pub fn run_on_simulated_device() -> Result<()> {
-    // Device initialization and engagement
-    let (engaged_state, qr_code_uri) = initialise_session()?;
-
-    // Reader processing QR and requesting needed fields
-    let (mut reader_session_manager, request) = establish_reader_session(qr_code_uri)?;
-
-    // Device accepting request
-    let (device_session_manager, requested_items) = handle_request(engaged_state, request)?;
-
-    // Prepare response with required elements
-    let response = create_response(
-        device_session_manager,
-        requested_items,
-        &create_signing_key()?,
-    )?;
-
-    // Reader Processing mDL data
-    reader_handle_device_response(&mut reader_session_manager, response)?;
-
-    Ok(())
-}
+#[allow(dead_code)]
+fn main() {}
 
 /// Parse the mDL encoded string into a [Documents] object.
-fn parse_mdl() -> Result<NonEmptyMap<DocType, Document>> {
+#[allow(dead_code)]
+pub fn parse_mdl() -> Result<NonEmptyMap<DocType, Document>> {
     let mdl_encoded = include_str!("data/stringified-mdl.txt");
     let mdl = Document::parse(mdl_encoded.to_string()).context("could not parse mDL")?;
-    let docs = Documents::new(DOC_TYPE.to_string(), mdl);
+    let docs = Documents::new(crate::DOC_TYPE.to_string(), mdl);
     Ok(docs)
 }
 
 /// Creates a QR code containing `DeviceEngagement` data, which includes its public key.
-fn initialise_session() -> Result<(SessionManagerEngaged, String)> {
+#[allow(dead_code)]
+pub fn initialise_session() -> Result<(SessionManagerEngaged, String)> {
     // Parse the mDL
     let docs = parse_mdl()?;
 
@@ -73,8 +47,22 @@ fn initialise_session() -> Result<(SessionManagerEngaged, String)> {
         .context("could not generate qr engagement")
 }
 
+/// Establishes the reader session from the given QR code and create request for needed elements.
+#[allow(dead_code)]
+pub fn establish_reader_session(qr: String) -> Result<(reader::SessionManager, Vec<u8>)> {
+    let requested_elements = Namespaces::new(
+        NAMESPACE.into(),
+        DataElements::new(AGE_OVER_21_ELEMENT.to_string(), false),
+    );
+    let (reader_sm, session_request, _ble_ident) =
+        reader::SessionManager::establish_session(qr, requested_elements)
+            .context("failed to establish reader session")?;
+    Ok((reader_sm, session_request))
+}
+
 /// The Device handles the request from the reader and advances the state.
-fn handle_request(
+#[allow(dead_code)]
+pub fn handle_request(
     state: SessionManagerEngaged,
     request: Vec<u8>,
 ) -> Result<(device::SessionManager, RequestedItems)> {
@@ -92,7 +80,8 @@ fn handle_request(
 }
 
 /// Prepare response with required elements.
-fn create_response(
+#[allow(dead_code)]
+pub fn create_response(
     mut session_manager: device::SessionManager,
     requested_items: RequestedItems,
     key: &p256::ecdsa::SigningKey,
@@ -116,6 +105,18 @@ fn create_response(
         .ok_or(anyhow!("cannot prepare response"))
 }
 
-fn create_signing_key() -> Result<p256::ecdsa::SigningKey> {
+/// Reader Processing mDL data.
+#[allow(dead_code)]
+pub fn reader_handle_device_response(
+    reader_sm: &mut reader::SessionManager,
+    response: Vec<u8>,
+) -> Result<()> {
+    let res = reader_sm.handle_response(&response)?;
+    println!("{:?}", res);
+    Ok(())
+}
+
+#[allow(dead_code)]
+pub fn create_signing_key() -> Result<p256::ecdsa::SigningKey> {
     Ok(p256::SecretKey::from_sec1_pem(include_str!("data/sec1.pem"))?.into())
 }
