+++ tests/data/stringified-mdl.txt
