+++ tests/on_simulated_reader.rs
@@ -1,6 +1,59 @@
-include!("../examples/on_simulated_reader.rs");
+mod common;
+
+use anyhow::{Context, Result};
+
+use isomdl::definitions::device_request::{DataElements, Namespaces};
+use isomdl::presentation::reader;
+
+use crate::common::{create_response, create_signing_key, handle_request, initialise_session};
+
+const DOC_TYPE: &str = "org.iso.18013.5.1.mDL";
+const NAMESPACE: &str = "org.iso.18013.5.1";
+const AGE_OVER_21_ELEMENT: &str = "age_over_21";
+
+fn main() {}
 
 #[test]
-fn on_simulated_reader() {
-    run_on_simulated_reader().unwrap();
+fn on_simulated_reader() -> Result<()> {
+    // Device initialization and engagement
+    let (engaged_state, qr_code_uri) = initialise_session()?;
+
+    // Reader processing QR and requesting needed fields
+    let (mut reader_session_manager, request) = establish_reader_session(qr_code_uri)?;
+
+    // Device accepting request
+    let (device_session_manager, requested_items) = handle_request(engaged_state, request)?;
+
+    // Prepare response with required elements
+    let response = create_response(
+        device_session_manager,
+        requested_items,
+        &create_signing_key()?,
+    )?;
+
+    // Reader Processing mDL data
+    reader_handle_device_response(&mut reader_session_manager, response)?;
+
+    Ok(())
+}
+/// Establishes the reader session from the given QR code and create request for needed elements.
+fn establish_reader_session(qr: String) -> Result<(reader::SessionManager, Vec<u8>)> {
+    let requested_elements = Namespaces::new(
+        NAMESPACE.into(),
+        DataElements::new(AGE_OVER_21_ELEMENT.to_string(), false),
+    );
+    let (reader_sm, session_request, _ble_ident) =
+        reader::SessionManager::establish_session(qr, requested_elements)
+            .context("failed to establish reader session")?;
+    Ok((reader_sm, session_request))
+}
+
+/// Reader Processing mDL data.
+fn reader_handle_device_response(
+    reader_sm: &mut reader::SessionManager,
+    response: Vec<u8>,
+) -> Result<()> {
+    let res = reader_sm.handle_response(&response)?;
+    println!("{:?}", res);
+    Ok(())
 }
